Consolidate PLP state: ActivePlpStream struct, split row module, allow target-type switch
- Replace active_plp_column / active_plp_target_type / active_plp_encoding
  with Option<ActivePlpStream { column, encoding }> in StmtState; the three
  scattered fields collapse into one cohesive struct.

- Remove the mid-stream target-type consistency check: msodbcsql allows the
  caller to switch SQL_C_CHAR / SQL_C_WCHAR between repeated SQLGetData calls
  on the same PLP column, so target_type is dropped from ActivePlpStream.

- Split row/mod.rs into row/mod.rs (re-export only) + row/row_writer.rs
  (OdbcRowWriter + PlpEncoding), matching the pattern used by other modules.
